+++ src/vec2.rs
@@ -798,4 +798,18 @@ impl<T> MapExt for Vec2<T> {
     }
 }
 
+impl<T: Default> FromIterator<T> for Vec2<T> {
+    fn from_iter<I: IntoIterator<Item = T>>(iter: I) -> Self {
+        let mut i = iter.into_iter();
+        Self::new(i.next().unwrap_or_default(), i.next().unwrap_or_default())
+    }
+}
+
+impl<'a, T: Default + Copy> FromIterator<&'a T> for Vec2<T> {
+    fn from_iter<I: IntoIterator<Item = &'a T>>(iter: I) -> Self {
+        let mut i = iter.into_iter().copied();
+        Self::new(i.next().unwrap_or_default(), i.next().unwrap_or_default())
+    }
+}
+
 impl_traits!(Vec2<T> [<Array, 2>, <Tuple, (x, y)>] => VecTraits);